Moving document and anchorable view models to their respective folders
