+++ CID_Tester/MainWindow.xaml
@@ -7,27 +7,32 @@
         xmlns:local="clr-namespace:CID_Tester"
         xmlns:vm="clr-namespace:CID_Tester.ViewModel"
         xmlns:StyleSelector="clr-namespace:CID_Tester.ViewModel.StyleSelectors"
+        
         xmlns:documents="clr-namespace:CID_Tester.View.Document"
+        xmlns:DocVm="clr-namespace:CID_Tester.ViewModel.Document"
+        
         xmlns:anchorable="clr-namespace:CID_Tester.View.Anchorables"
-        xmlns:controlsVm="clr-namespace:CID_Tester.ViewModel.Controls"
+        xmlns:AnchVm="clr-namespace:CID_Tester.ViewModel.Anchorables"
+        
+        xmlns:controls="clr-namespace:CID_Tester.View.Controls"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800"
         WindowState="Maximized"
         UseLayoutRounding="True">
     <Window.Resources>
-        <DataTemplate DataType="{x:Type vm:DashboardViewModel}">
+        <DataTemplate DataType="{x:Type DocVm:DashboardViewModel}">
             <documents:DashboardView />
         </DataTemplate>
-        <DataTemplate DataType="{x:Type vm:DevicesViewModel}">
+        <DataTemplate DataType="{x:Type DocVm:DevicesViewModel}">
             <documents:DevicesView />
         </DataTemplate>
-        <DataTemplate DataType="{x:Type vm:TestPlanViewModel}">
+        <DataTemplate DataType="{x:Type DocVm:TestPlanViewModel}">
             <documents:TestPlanView />
         </DataTemplate>
-        <DataTemplate DataType="{x:Type vm:ResultsViewModel}">
+        <DataTemplate DataType="{x:Type DocVm:ResultsViewModel}">
             <documents:ResultsView />
         </DataTemplate>
-        <DataTemplate DataType="{x:Type controlsVm:TestParameterPropertiesViewModel}">
+        <DataTemplate DataType="{x:Type AnchVm:TestParameterPropertiesViewModel}">
             <anchorable:TestParameterProperties />
         </DataTemplate>
